Shifted faction select to top-level
