+++ src/components/CharacterHeader.css
@@ -27,19 +27,6 @@
   overflow: hidden;
 }
 
-.characterHeader .faction-label {
-  grid-column: 1;
-  grid-row: 1 / 2;
-  align-self: center;
-  border: none;
-}
-
-.characterHeader .faction-select {
-  grid-column: 1 / 3;
-  grid-row: 1/2;
-  align-self: center;
-}
-
 .characterHeader .unit-type-select {
   grid-column: 3 / 4;
   grid-row: 1 / 2;
